Versioning Logic
-Added the versioning logic(NOT IMPLEMENTED)
- I didn't notice where the edit page is maybe due to the night hours so I wasn't sure where to insert the create_quotation_version and stuff but the versioning logic I think is complete(?) someone try going over that when possible.

For displaying the quotations, I think we can either show all but specifically try to sort the rows somehow to show the active versions first then by project the other version that are inactive. Assuming the versioning works we have that but are lacking a way to properly view the inactive if they are needed for whatever reason and not having them sorted would be confusing as well.